Rewrite all widgets to make it more configurable for implementing scrollarea
