+++ src/components/Image/index.ts
@@ -1,43 +1,9 @@
-import {
-  QPixmap,
-  QLabelEvents,
-  AspectRatioMode,
-  NodeWidget
-} from "@nodegui/nodegui";
+import { QLabelEvents, NodeWidget } from "@nodegui/nodegui";
 import { Fiber } from "react-reconciler";
 import { registerComponent, ComponentConfig } from "../config";
-import { ImageLabel } from "./ImageLabel";
-import { TextProps, setProps as setTextProps } from "../Text";
-interface ImageProps extends TextProps {
-  src?: string;
-  aspectRatioMode?: AspectRatioMode;
-}
-
-const setProps = (
-  widget: ImageLabel,
-  newProps: ImageProps,
-  oldProps: ImageProps
-) => {
-  const setter: ImageProps = {
-    set src(imageUrl: string) {
-      if (!imageUrl) {
-        return;
-      }
-      const pixMap = new QPixmap(imageUrl);
-      widget.setPixmap(pixMap);
-      const size = widget.size();
-      widget.scalePixmap(size.width, size.height);
-    },
-    set aspectRatioMode(mode: AspectRatioMode) {
-      widget.setAspectRatioMode(mode);
-    }
-  };
-  Object.assign(setter, newProps);
-  setTextProps(widget, newProps, oldProps);
-};
-
+import { RNImage, setProps, ImageProps } from "./RNImage";
 class ImageConfig extends ComponentConfig {
-  id = "image";
+  tagName = RNImage.tagName;
   shouldSetTextContent(nextProps: object): boolean {
     return true;
   }
@@ -47,7 +13,7 @@ class ImageConfig extends ComponentConfig {
     context: any,
     workInProgress: Fiber
   ): NodeWidget {
-    const widget = new ImageLabel();
+    const widget = new RNImage();
     setProps(widget, newProps, {});
     widget.addEventListener(QLabelEvents.Resize, () => {
       const size = widget.size();
@@ -62,7 +28,7 @@ class ImageConfig extends ComponentConfig {
     newProps: object,
     finishedWork: Fiber
   ): void {
-    setProps(instance as ImageLabel, newProps, oldProps);
+    setProps(instance as RNImage, newProps, oldProps);
   }
 }
