+++ src/components/PlainTextEdit/index.ts
@@ -1,33 +1,13 @@
-import { QPlainTextEdit, NodeWidget } from "@nodegui/nodegui";
-import { ViewProps, setProps as setViewProps } from "../View";
+import { NodeWidget } from "@nodegui/nodegui";
 import { registerComponent, ComponentConfig } from "../config";
 import { Fiber } from "react-reconciler";
-
-interface PlainTextEditProps extends ViewProps {
-  children?: string;
-  text?: string;
-  readOnly?: boolean;
-}
-
-const setProps = (
-  widget: QPlainTextEdit,
-  newProps: PlainTextEditProps,
-  oldProps: PlainTextEditProps
-) => {
-  const setter: PlainTextEditProps = {
-    set text(text: string) {
-      text ? widget.setPlainText(text) : widget.clear();
-    },
-    set readOnly(isReadOnly: boolean) {
-      widget.setReadOnly(isReadOnly);
-    }
-  };
-  Object.assign(setter, newProps);
-  setViewProps(widget, newProps, oldProps);
-};
-
+import {
+  RNPlainTextEdit,
+  setProps,
+  PlainTextEditProps
+} from "./RNPlainTextEdit";
 class PlainTextEditConfig extends ComponentConfig {
-  id = "plaintextedit";
+  tagName = RNPlainTextEdit.tagName;
   shouldSetTextContent(nextProps: object): boolean {
     return true;
   }
@@ -37,7 +17,7 @@ class PlainTextEditConfig extends ComponentConfig {
     context: any,
     workInProgress: Fiber
   ): NodeWidget {
-    const widget = new QPlainTextEdit();
+    const widget = new RNPlainTextEdit();
     setProps(widget, newProps, {});
     return widget;
   }
@@ -48,7 +28,7 @@ class PlainTextEditConfig extends ComponentConfig {
     newProps: object,
     finishedWork: Fiber
   ): void {
-    setProps(instance as QPlainTextEdit, newProps, oldProps);
+    setProps(instance as RNPlainTextEdit, newProps, oldProps);
   }
 }
