+++ src/components/ScrollArea/index.ts
@@ -1,30 +1,10 @@
 import { registerComponent, ComponentConfig } from "../config";
 import { QScrollArea, NodeWidget } from "@nodegui/nodegui";
-import { ViewProps, setProps as setViewProps } from "../View";
-import { ReactElement } from "react";
 import { Fiber } from "react-reconciler";
-
-export interface ScrollAreaProps extends ViewProps {
-  children?: ReactElement;
-}
-
-export const setProps = (
-  widget: QScrollArea,
-  newProps: ScrollAreaProps,
-  oldProps: ScrollAreaProps
-) => {
-  const setter: ScrollAreaProps = {
-    set children(childWidget: ReactElement) {
-      console.log(childWidget, "childWidget");
-      // widget.setWidget(childWidget);
-    }
-  };
-  Object.assign(setter, newProps);
-  setViewProps(widget, newProps, oldProps);
-};
+import { RNScrollArea, setProps, ScrollAreaProps } from "./RNScrollArea";
 
 class ScrollAreaConfig extends ComponentConfig {
-  id = "scrollarea";
+  tagName = RNScrollArea.tagName;
   shouldSetTextContent(nextProps: object): boolean {
     return false;
   }
@@ -34,7 +14,7 @@ class ScrollAreaConfig extends ComponentConfig {
     context: any,
     workInProgress: Fiber
   ): NodeWidget {
-    const widget = new QScrollArea();
+    const widget = new RNScrollArea();
     setProps(widget, newProps, {});
     return widget;
   }
