+++ src/components/Text/index.ts
@@ -1,31 +1,9 @@
-import { QLabel, NodeWidget } from "@nodegui/nodegui";
+import { NodeWidget } from "@nodegui/nodegui";
 import { Fiber } from "react-reconciler";
-import { ViewProps, setProps as setViewProps } from "../View";
 import { registerComponent, ComponentConfig } from "../config";
-export interface TextProps extends ViewProps {
-  children?: string | number;
-  wordWrap?: boolean;
-}
-
-export const setProps = (
-  widget: QLabel,
-  newProps: TextProps,
-  oldProps: TextProps
-) => {
-  const setter: TextProps = {
-    set children(text: string | number) {
-      widget.setText(text);
-    },
-    set wordWrap(shouldWrap: boolean) {
-      widget.setWordWrap(shouldWrap);
-    }
-  };
-  Object.assign(setter, newProps);
-  setViewProps(widget, newProps, oldProps);
-};
-
+import { RNText, TextProps, setProps } from "./RNText";
 class TextConfig extends ComponentConfig {
-  id = "text";
+  tagName = RNText.tagName;
   shouldSetTextContent(nextProps: object): boolean {
     return true;
   }
@@ -35,7 +13,7 @@ class TextConfig extends ComponentConfig {
     context: any,
     workInProgress: Fiber
   ): NodeWidget {
-    const widget = new QLabel();
+    const widget = new RNText();
     setProps(widget, newProps, {});
     return widget;
   }
@@ -46,7 +24,7 @@ class TextConfig extends ComponentConfig {
     newProps: object,
     finishedWork: Fiber
   ): void {
-    setProps(instance as QLabel, newProps, oldProps);
+    setProps(instance as RNText, newProps, oldProps);
   }
 }
