+++ src/components/Window/index.ts
@@ -1,38 +1,10 @@
 import { registerComponent, ComponentConfig } from "../config";
-import { QMainWindow, QWidget, FlexLayout, NodeWidget } from "@nodegui/nodegui";
-import { ViewProps, setProps as setViewProps } from "../View";
+import { QWidget, FlexLayout, NodeWidget } from "@nodegui/nodegui";
 import { Fiber } from "react-reconciler";
-interface WindowProps extends ViewProps {
-  centralWidgetProps?: ViewProps;
-}
-
-const setProps = (
-  window: QMainWindow,
-  newProps: WindowProps,
-  oldProps: WindowProps
-) => {
-  const setter: WindowProps = {
-    set centralWidgetProps(centralWidgetProps: object) {
-      if (window.centralWidget) {
-        const oldcentralWidgetProps = oldProps.centralWidgetProps || {};
-        setViewProps(
-          window.centralWidget,
-          centralWidgetProps,
-          oldcentralWidgetProps
-        );
-      } else {
-        console.warn(
-          "Trying to set viewProps for main window but no central widget set."
-        );
-      }
-    }
-  };
-  Object.assign(setter, newProps);
-  setViewProps(window, newProps, oldProps);
-};
+import { WindowProps, setProps, RNWindow } from "./RNWindow";
 
 class WindowConfig extends ComponentConfig {
-  id = "window";
+  tagName = RNWindow.tagName;
   shouldSetTextContent(nextProps: object): boolean {
     return false;
   }
@@ -42,7 +14,7 @@ class WindowConfig extends ComponentConfig {
     context: any,
     workInProgress: Fiber
   ): NodeWidget {
-    const window = new QMainWindow();
+    const window = new RNWindow();
     const rootView = new QWidget();
     const rootViewLayout = new FlexLayout();
     rootViewLayout.setFlexNode(rootView.getFlexNode());
@@ -76,7 +48,7 @@ class WindowConfig extends ComponentConfig {
     newProps: object,
     finishedWork: Fiber
   ): void {
-    setProps(instance as QMainWindow, newProps, oldProps);
+    setProps(instance as RNWindow, newProps, oldProps);
   }
 }
